add first unit test
Not sure if I like this. It feels quite heavy handed, but I guess this
is just the initial scaffold setup.

Testicate depends heavily on the vscode object. Create a mock of this in
the tests and inject it in. This works, but I don't like it. It means
the tests know a lot about the internal workings of the code.

I also stole the InMemoryDocument class from the vscode core markdown
extension - this is used in the mock.

Move to two space indents - since we are using that in the main
extension.